Add another container and a docker-compose yaml
Flask does not come with a deployment http server I added Nginx which
will serve as a reverse proxy and will pass the request to the gateway
interface which will pass them to Flask. The web server is put in a
seperate container and configured to serve the applycation. The containers
are composed with a docker-compose.yml file and now port 5000 which is
used by Flask is only internally exposed for Nginx to pass the requests to.