Add a Docker container for running the tests
A Dockerfile has been added to allow the tests to be run in an isolated
environment. The Dockerfile uses the requirements for the web application
container but adds pytest to then. A Bash script to build and run the
container was added. The Bash script can be run from either the root
project folder or the tests folder.