Modify the package to be executable in Docker
The Python package was previously wrongly configured. The MANIFEST.in
and README.rst have been added to fix the non-code files missing from
the package. Then the files were modified to run the web socket
gateway interface instead of the flask app. The Dockerfile is modified
to run the python package generateme instead of the files in the folder.