Update the image method to not use files
Previously the image generation saved the new image then the website
would open the image and present it but now it is generating it and
converting to a numpy array and then a PIL image and sending to the
website to be presented directly. This eliminates the change of error
when multiple users are simultaniously using the website.