Add json api for generated images
Three new endpoints have been added for each of the GANs. There
endpoints return images in a Json format. The Json has a completion
status that states if the image was delived and an image attribute
which contains an image converted to BytesIO then encoded to Base64
then endoded as UTF-8 to be a string.